requirements.txt
MouseInfo==0.1.3
numpy==1.18.2
Pillow==7.1.1
PyAutoGUI==0.9.50
pyclick==0.0.2
pygame==1.9.6
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.3
python3-xlib==0.15
PyTweening==1.0.3
pywin32==227
six==1.14.0
treys==0.1.3
xlib==0.21
